BookStoreTests/Controllers/BooksControllerTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using BookStore.Controllers;
using System;
using System.Collections.Generic;
using System.Text;
using Microsoft.AspNetCore.Mvc;
using BookStore.Services;

namespace BookStore.Controllers.Tests
{

    [TestClass()]
    public class BooksControllerTests
    {
        private FakeBookService bookService = new FakeBookService();

        //public BooksController controller = new BooksController(bookService, categoryService, publisherService, writerService);

        [TestMethod()]
        public void BooksControllerTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void IndexTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void CreateTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void CreateTest1()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void EditTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void EditTest1()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void DeleteTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void DeleteOkTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void DetailsTest()
        {
            var controller = new BooksController();
            var result = controller.Details(2) as ViewResult;
            Assert.AreEqual("Details", result.ViewName);

        }
    }
}